GameOfLife, IsGraphBipartite, PalindromeNumber, ArithmeticPower, GraphDivideTwoSetsOrBipartiteGraph

+++ Java/companyWise/facebook/ArithmeticPower.java
@@ -20,6 +20,8 @@ public class ArithmeticPower {
         System.out.printf("\n-2^(-3) = " + ArithmeticPowerSimple.arithmeticPower(-2, 3));
         System.out.printf("\n-2^(-3) = " + ArithmeticPowerRecursive.arithmeticPower(-2, 3));
         System.out.printf("\n-2^(-3) = " + ArithmeticPowerIterative.arithmeticPower(-2, 3));
+
+
     }
 }
 
@@ -32,7 +34,7 @@ class ArithmeticPowerSimple {
     }
 
 
-    private static long power(int x, int y) {
+    private static double power(double x, int y) {
         if (y == 0)
             return 1;
         else if (y % 2 == 0)
@@ -44,12 +46,12 @@ class ArithmeticPowerSimple {
 
 //O(logy) / O(logy)
 class ArithmeticPowerRecursive {
-    public static double arithmeticPower(int x, int y) {
+    public static double arithmeticPower(double x, int y) {
 
         return (y > 0) ? power(x, Math.abs(y)) : (double) 1 / power(x, Math.abs(y));
     }
 
-    private static long power(int x, int y) {
+    private static double power(double x, int y) {
         if (y == 0)
             return 1;
         else if (y % 2 == 0)
@@ -61,14 +63,14 @@ class ArithmeticPowerRecursive {
 
 //O(logy)/O(1)
 class ArithmeticPowerIterative {
-    public static double arithmeticPower(int x, int y) {
+    public static double arithmeticPower(double x, int y) {
 
         return (y > 0) ? power(x, Math.abs(y)) : (double) 1 / power(x, Math.abs(y));
     }
 
-    private static int power(int x, int y) {
+    private static double power(double x, int y) {
         // Initialize result
-        int res = 1;
+        double res = 1;
 
         while (y > 0) {
             // If y is odd,
